// website-next/src/lib/client/types.gen.ts
export type KillDistributionPointResponse = {
    bucketStart?: string;
    kills?: number;
};